refactor: refactor import paths in SignInPage.tsx, SignUpPage.tsx, calenderConstants.ts, App.tsx, parseQueryArgs.ts, SignUpEmailPasswordForm.tsx, SignUpValidationForm.tsx, routes.ts, main.tsx, CalenderDay.styles.ts, tsconfig.json, vite.config.ts, useSuspendedQuery.ts, userQueries.ts
