+++ src/constants/calenderConstants.ts
@@ -1,3 +1,3 @@
-import { DAYS } from '../grTypes/calender/internal/internalCalnderTypes.ts';
+import { DAYS } from '@/grTypes/calender/internal/internalCalnderTypes.ts';
 
 export const WEEKDAYS: DAYS[] = ['Sun', 'Mon', 'Tue', 'Wed', 'Thu', 'Fri', 'Sat'];
