+++ src/constants/routes.ts
@@ -2,6 +2,7 @@ const ROUTES = {
   MAIN_PAGE: '/',
   SIGN_UP: '/signup',
   SIGN_IN: '/signin',
+  CALENDER: '/calender',
 };
 
 export default ROUTES;
